@@ -9,7 +9,7 @@ import '../widgets/print_receipt.dart';
 
 class CheckIn extends StatefulWidget {
   static const name = "check-in";
-  const CheckIn({Key? key}) : super(key: key);
+  const CheckIn({super.key});
 
   @override
   State<CheckIn> createState() => _CheckInState();
@@ -100,6 +100,8 @@ class _CheckInState extends State<CheckIn> {
                               items: List.generate(
                                 response.data!["vehical_type"].length,
                                 (index) => DropdownMenuItem(
+                                  value: response.data!["vehical_type"][index]
+                                      .toString(),
                                   child: Row(
                                     children: [
                                       Text(response.data!["vehical_type"][index]
@@ -109,8 +111,6 @@ class _CheckInState extends State<CheckIn> {
                                           ["rate"]),
                                     ],
                                   ),
-                                  value: response.data!["vehical_type"][index]
-                                      .toString(),
                                 ),
                               ),
                               onChanged: (v) {
@@ -159,6 +159,8 @@ class _CheckInState extends State<CheckIn> {
                                       response.data!["province"].length,
                                       // response.data!["vehical_type"].length,
                                       (index) => DropdownMenuItem(
+                                        value: response.data!["province"][index]
+                                            .toString(),
                                         child: Row(
                                           mainAxisAlignment:
                                               MainAxisAlignment.start,
@@ -172,8 +174,6 @@ class _CheckInState extends State<CheckIn> {
                                             //     ["code"]),
                                           ],
                                         ),
-                                        value: response.data!["province"][index]
-                                            .toString(),
                                       ),
                                     ),
                                     onChanged: (v) {
@@ -220,6 +220,8 @@ class _CheckInState extends State<CheckIn> {
                                       response.data!["zone"].length,
                                       // response.data!["vehical_type"].length,
                                       (index) => DropdownMenuItem(
+                                        value: response.data!["zone"][index]
+                                            .toString(),
                                         child: Row(
                                           mainAxisAlignment:
                                               MainAxisAlignment.start,
@@ -233,8 +235,6 @@ class _CheckInState extends State<CheckIn> {
                                             //     ["code"]),
                                           ],
                                         ),
-                                        value: response.data!["zone"][index]
-                                            .toString(),
                                       ),
                                     ),
                                     onChanged: (v) {
