@@ -8,7 +8,7 @@ import 'package:smartparking/screens/root_screen.dart';
 
 class SplashScreen extends StatefulWidget {
   static const name = "splash";
-  const SplashScreen({Key? key}) : super(key: key);
+  const SplashScreen({super.key});
 
   @override
   State<SplashScreen> createState() => _SplashScreenState();
